decmail请求参数报错——可能是json。

+++ ConAppDotNettyRPCTest.Client/IOrderService.cs
@@ -6,6 +6,6 @@ namespace ConAppDotNettyRPCTest.Client
 {
     public interface IOrderService
     {
-        double CalculateFinalOrderSum(long userId, double originalSum);
+        decimal CalculateFinalOrderSum(long userId, decimal originalSum);
     }
 }
